+++ apps/src/templates/certificates/petition/PetitionCallToAction.story.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import PetitionCallToAction from '@cdo/apps/templates/certificates/PetitionCallToAction';
+import PetitionCallToAction from '@cdo/apps/templates/certificates/petition/PetitionCallToAction';
 
 export default storybook => {
   return storybook
